// types/type.d.ts
export declare interface OilmillTransaction {
  booking: {
    deliveryDate: string;
    description: string;
    estimatedWeight: number;
    plateNumber: string;
  };
  bookingId: string;
  copraOwner: {
    email: string;
    name: string;
  };
  actualWeight: number | null;
  qualityGrade: string | null;
  copraOwnerEmail: string;
  copraOwnerId: string;
  copraOwnerName: string;
  date: string;
  dateTrans: string;
  dateUpdate: string;
  discount: number | null;
  id: string;
  invoiceID: string | null;
  oilMillCompanyId: string;
  paymentType: string;
  payoutID: string | null;
  plateNumber: string;
  remarks: string | null;
  status: string;
  time: string;
  totalAmount: number | null;
}

export declare interface CopraOwnerTransaction {
  booking: {
    deliveryDate: string;
    description: string;
    estimatedWeight: number;
  };
  bookingId: string;
  copraOwnerId: string;
  date: string;
  dateTrans: string;
  dateUpdate: string;
  discount: number | null;
  id: string;
  invoiceID: string | null;
  oilMillCompany: {
    address: string;
    name: string;
  };
  actualWeight: number | null;
  qualityGrade: string | null;
  oilMillCompanyAddress: string;
  oilMillCompanyId: string;
  oilMillCompanyName: string;
  paymentType: string;
  payoutID: string | null;
  plateNumber: string;
  remarks: string | null;
  status: string;
  time: string;
  totalAmount: number | null;
}
// ...
export interface PaymentModalProps {
  visible: boolean;
  transaction: CopraOwnerTransaction | OilmillTransaction | null;
  onConfirm: () => void;
  onClose: () => void;
}